labels fade out and then get free'd
ideally, they would go back to a pool. But i'm not going to do that ATM.
Also, ideally the animation player would call a method, however I'm getting `Method not found in object: OnFadeComplete` from the editor when i choose the method ... even though it exists. and i know godot knows it exists, because it shows the method